Convert all src/codeql-cli JavaScript files to TypeScript (#57369)
